# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for Hotel Reservation System (GoTel).",
        "title": "GoTel API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ramin Fathi",
            "email": "support@gotel.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api/v1",
    "paths": {
        "/admin/booking": {
            "get": {
                "description": "Get a list of all bookings in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get all bookings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Booking"
                            }
                        }
                    }
                }
            }
        },
        "/admin/hotel": {
            "post": {
                "description": "Add a new hotel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Add a hotel",
                "parameters": [
                    {
                        "description": "Hotel Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateHotelParams"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Hotel"
                        }
                    }
                }
            }
        },
        "/admin/hotel/{id}": {
            "put": {
                "description": "Update hotel details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update a hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateHotelParams"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/booking": {
            "get": {
                "description": "Get all bookings for the logged-in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Get my bookings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Booking"
                            }
                        }
                    }
                }
            }
        },
        "/booking/{id}": {
            "get": {
                "description": "Get a single booking by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Get booking details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Booking"
                        }
                    }
                }
            }
        },
        "/booking/{id}/cancel": {
            "post": {
                "description": "Cancel a booking (Method is POST/PUT for safety)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Cancel booking",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hotel": {
            "get": {
                "description": "Get a list of all hotels with filtering options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotel"
                ],
                "summary": "Get all hotels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Hotel"
                            }
                        }
                    }
                }
            }
        },
        "/hotel/{id}": {
            "get": {
                "description": "Get detailed information of a specific hotel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotel"
                ],
                "summary": "Get hotel by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Hotel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/hotel/{id}/rooms": {
            "get": {
                "description": "Get all rooms belonging to a specific hotel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hotel"
                ],
                "summary": "Get hotel rooms",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "X-Api-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Room"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.Booking": {
            "type": "object",
            "properties": {
                "canceled": {
                    "type": "boolean"
                },
                "fromDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "numPersons": {
                    "type": "integer"
                },
                "roomID": {
                    "type": "string"
                },
                "tillDate": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "types.CreateHotelParams": {
            "type": "object",
            "required": [
                "location",
                "name"
            ],
            "properties": {
                "location": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
...
        "types.Hotel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "types.Room": {
            "type": "object",
            "properties": {
                "basePrice": {
                    "type": "number"
                },
                "hotelId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/types.RoomType"
                }
            }
        },
        "types.RoomType": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "Single",
                "Double",
                "SeaView",
                "KingSuite"
            ]
        },
        "types.UpdateHotelParams": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "name": {
                    "description": "omitempty یعنی اگر نفرستاد گیر نده",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Api-Token",
            "in": "header"
        }
    }
}